Added new function to simultaneously activate and override purchase
I suspect that the purchase activation on new users with overrides was
messing up, because of the observation callbacks. Because the overrides
and the activation changes were being done asynchronously, it may have
been overriding and then inactivating the purchase. So I did it in the
same mongo update in order to alleviate that potential problem